10883 - moved category section's dsm into CategoriesDsm.jsx
